Split component and add localisation
